tests/requirements.txt
docker-py==1.2.3
requests==2.5.3
lxml==3.4.4